App Delegate Logic & Brunch Countdown
Added logic to the app delegate that checks to see if the dayOfTheWeek
is Saturday (7) or Sunday (1). If yes it sends you off to the Get
Started Storyboard. If not it sends you off to the Closed Storyboard.

On the Closed Storyboard I added logic that compares the current date &
time to the next upcoming brunch date & time. It’s currently un-styled
but the countdown timer displays a simple “XX to go” message depending
on how many days, hours, or minutes are left to go before it’s time for
brunch.